+++ src/layer/riscv/gru_riscv_zfh.cpp
@@ -138,8 +138,8 @@ static int gru_fp16s(const Mat& bottom_blob, Mat& top_blob, int reverse, const M
 
             // sigmoid(R)
             // sigmoid(U)
-            R = 1.f / (1.f + exp(-R));
-            U = 1.f / (1.f + exp(-U));
+            R = 1.f / (1.f + expf(-R));
+            U = 1.f / (1.f + expf(-U));
 
             // gate new
             const float* bias_c_WN = bias_c.row(2);
@@ -215,7 +215,7 @@ static int gru_fp16s(const Mat& bottom_blob, Mat& top_blob, int reverse, const M
 #endif // __riscv_zvfh
 
             // tanh(N)
-            N = tanh(N);
+            N = tanhf(N);
 
             gates_data[0] = U;
             gates_data[1] = N;
@@ -502,8 +502,8 @@ static int gru_fp16sa(const Mat& bottom_blob, Mat& top_blob, int reverse, const 
 
             // sigmoid(R)
             // sigmoid(U)
-            R = 1.f / (1.f + (__fp16)exp((float)(-R)));
-            U = 1.f / (1.f + (__fp16)exp((float)(-U)));
+            R = 1.f / (1.f + (__fp16)expf((float)(-R)));
+            U = 1.f / (1.f + (__fp16)expf((float)(-U)));
 
             // gate new
             const __fp16* bias_c_WN = bias_c.row<const __fp16>(2);
@@ -575,7 +575,7 @@ static int gru_fp16sa(const Mat& bottom_blob, Mat& top_blob, int reverse, const 
 #endif // __riscv_zvfh
 
             // tanh(N)
-            N = (__fp16)tanh((float)N);
+            N = (__fp16)tanhf((float)N);
 
             gates_data[0] = U;
             gates_data[1] = N;
