+++ src/layer/riscv/rvv_mathfun.h
@@ -568,7 +568,7 @@ _RVV_FLOAT32_ERFC_OP(8, 4)
         __riscv_vse32_v_f32m##LMUL(tmpy.data(), b, vl);                                                          \
         for (size_t i = 0; i < vl; i++)                                                                          \
         {                                                                                                        \
-            tmpx[i] = atan2(tmpx[i], tmpy[i]);                                                                   \
+            tmpx[i] = atan2f(tmpx[i], tmpy[i]);                                                                   \
         }                                                                                                        \
         return __riscv_vle32_v_f32m##LMUL(tmpx.data(), vl);                                                      \
     }
